A little refactor to consolidate terminal state functions
