+++ pkg/multiterm/bufferedterm_test.go
@@ -2,13 +2,7 @@ package multiterm
 
 import (
 	"testing"
-
-	"github.com/stretchr/testify/assert"
 )
-
-func TestIsPipedOutput(t *testing.T) {
-	assert.True(t, IsPipedOutput())
-}
 
 func TestBufferedTerm(t *testing.T) {
 	vt := NewBufferedTerm()
